move ai response parser
