+++ services/storyteller/index.ts
@@ -6,7 +6,7 @@
 export * from './api';
 export * from '../dialogueService';
 export * from "./systemPrompt";
-export * from '../aiResponseParser';
+export * from './responseParser';
 export * from '../mapUpdateService';
 export * from '../modelDispatcher';
 export * from '../geminiClient';
